+++ cmd/main_synther_not_native.go
@@ -1,7 +1,0 @@
-//go:build !native
-
-package cmd
-
-import "github.com/vsariola/sointu/vm"
-
-var MainSynther = vm.GoSynther{}