+++ cmd/sointu-vsti/nameid_native.go
@@ -1,6 +1,0 @@
-//go:build native
-
-package main
-
-var PLUGIN_ID = [4]byte{'S', 'n', 't', 'N'}
-var PLUGIN_NAME = "Sointu Native"