+++ cmd/synthers.go
@@ -1,0 +1,10 @@
+package cmd
+
+import (
+	"github.com/vsariola/sointu"
+	"github.com/vsariola/sointu/vm"
+)
+
+var Synthers = []sointu.Synther{
+	vm.GoSynther{},
+}